fix: resolve high-severity audit issues (H1, H3, H4, H5, H6)
- H1: tighten dependency constraints (activemodel >= 6.1 < 9, nokogiri >= 1.13, ruby >= 3.1)
- H3: extract SEPA::Address base class, DRY creditor/debtor address + build_postal_address
- H4: typed error hierarchy (ValidationError, SchemaValidationError)
- H5: skip empty PmtTpInf when no service_level or category_purpose
- H6: enforce SEPA max amount 999_999_999.99
