+++ lib/sepa_king.rb
@@ -8,6 +8,7 @@ require 'sepa_king/converter'
 require 'sepa_king/validator'
 require 'sepa_king/account'
 require 'sepa_king/account/debtor_account'
+require 'sepa_king/account/address'
 require 'sepa_king/account/debtor_address'
 require 'sepa_king/account/creditor_account'
 require 'sepa_king/account/creditor_address'
