+++ lib/sepa_king/error.rb
@@ -1,4 +1,5 @@
 module SEPA
-  class Error < RuntimeError
-  end
+  class Error < RuntimeError; end
+  class ValidationError < Error; end
+  class SchemaValidationError < Error; end
 end
