+++ lib/sepa_king/message/direct_debit.rb
@@ -160,44 +160,7 @@ module SEPA
         end
         builder.Dbtr do
           builder.Nm(transaction.name)
-          if transaction.debtor_address
-            builder.PstlAdr do
-              # Only set the fields that are actually provided.
-              # StrtNm, BldgNb, PstCd, TwnNm provide a structured address
-              # separated into its individual fields.
-              # AdrLine provides the address in free format text.
-              # Both are currently allowed and the actual preference depends on the bank.
-              # Also the fields that are required legally may vary depending on the country
-              # or change over time.
-              if transaction.debtor_address.street_name
-                builder.StrtNm transaction.debtor_address.street_name
-              end
-
-              if transaction.debtor_address.building_number
-                builder.BldgNb transaction.debtor_address.building_number
-              end
-
-              if transaction.debtor_address.post_code
-                builder.PstCd transaction.debtor_address.post_code
-              end
-
-              if transaction.debtor_address.town_name
-                builder.TwnNm transaction.debtor_address.town_name
-              end
-
-              if transaction.debtor_address.country_code
-                builder.Ctry transaction.debtor_address.country_code
-              end
-
-              if transaction.debtor_address.address_line1
-                builder.AdrLine transaction.debtor_address.address_line1
-              end
-
-              if transaction.debtor_address.address_line2
-                builder.AdrLine transaction.debtor_address.address_line2
-              end
-            end
-          end
+          build_postal_address(builder, transaction.debtor_address) if transaction.debtor_address
         end
         builder.DbtrAcct do
           builder.Id do
