+++ sepa_king.gemspec
@@ -16,10 +16,10 @@ Gem::Specification.new do |s|
   s.files         = `git ls-files`.split($/)
   s.require_paths = ['lib']
 
-  s.required_ruby_version = '>= 2.7'
+  s.required_ruby_version = '>= 3.1'
 
-  s.add_runtime_dependency 'activemodel', '>= 4.2'
-  s.add_runtime_dependency 'nokogiri'
+  s.add_runtime_dependency 'activemodel', '>= 6.1', '< 9'
+  s.add_runtime_dependency 'nokogiri', '>= 1.13'
   s.add_runtime_dependency 'iban-tools'
 
   s.add_development_dependency 'bundler'
